Logging, base class work, metadata
Summary:
Starting to add logging code so that the library can log debug level actions.

Extended the base class with some basic methods that need to return None or
similar so that calls to them will indicate that a specific driver isn't in
use.

The display is now initialised to None so that it's possible to detect when
the display is off/unplugged. Doesn't work quite right yet.

Test Plan:
Sample code doesn't work if the library doesn't work.